+++ tests/rmq-node.lisp
@@ -8,370 +8,251 @@
 (defparameter *source-queue* (format nil "test-source-~d" (get-universal-time)))
 (defparameter *dest-queue* (format nil "test-dest-~d" (get-universal-time)))
 (defparameter *fail-queue* (format nil "test-fail-~d" (get-universal-time)))
-(defparameter *conn* (let ((conn (setup-connection :host (uiop:getenv "TEST_RMQ"))))
-                       (if (getf conn :success)
-                           (getf conn :conn)
-                           (error (getf conn :error)))))
 (defvar *node* (make-rmq-node nil (format nil "test-rmq-node-~d" (get-universal-time))
-                              *conn* 1 *source-queue* *dest-queue* *fail-queue*))
+                              *source-queue* *dest-queue* *fail-queue*
+                              :host (uiop:getenv "TEST_RMQ")))
+(defvar *checking-node*
+  (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
+                 *dest-queue* *source-queue* *fail-queue*
+                 :batch-size 10 :host (uiop:getenv "TEST_RMQ")))
 (startup *node* nil)
+(startup *checking-node* nil)
 
 (subtest "test-full-message-path"
   (let ((test-msg (format nil "test-~d" (get-universal-time))))
-    (ok (getf (send-message *node* *source-queue* test-msg) :success))
+    (is (send-message *checking-node* *source-queue* test-msg) :amqp-status-ok)
     (sleep .1)
-    (let* ((got-msg (get-message *node*))
-           (inner-msg (getf got-msg :result)))
-      (ok (getf got-msg :success))
-      (ok (not (getf got-msg :timeout)))
-      (is (rmq-message-body inner-msg) test-msg)
-      (ok (getf (ack-message *node* inner-msg) :success)))))
+    (let ((got-msg (get-message *node*)))
+      (is (rmq-message-body got-msg) test-msg)
+      (ack-message *node* got-msg))))
 
-(subtest "get-message-handles-timeout"
-  (let ((got-msg (get-message *node*)))
-    (ok (getf got-msg :success))
-    (ok (getf got-msg :timeout))))
+(subtest "get-message-timeout"
+  (is-error (get-message *node*) 'rabbitmq-library-error))
 
 (subtest "nack works as expected (requeue)"
   (let ((test-msg (format nil "test-~d" (get-universal-time))))
-    (ok (getf (send-message *node* *source-queue* test-msg) :success))
+    (is (send-message *checking-node* *source-queue* test-msg) :amqp-status-ok)
     (sleep .1)
-    (let* ((got-msg (get-message *node*))
-           (inner-msg (getf got-msg :result)))
-      (ok (getf got-msg :success))
-      (ok (not (getf got-msg :timeout)))
-      (is (rmq-message-body inner-msg) test-msg)
-      (ok (getf (nack-message *node* inner-msg t) :success)))
+    (let ((got-msg (get-message *node*)))
+      (is (rmq-message-body got-msg) test-msg)
+      (is (nack-message *node* got-msg t) :amqp-status-ok))
     (sleep .1)
-    (let* ((got-msg (get-message *node*))
-           (inner-msg (getf got-msg :result)))
-      (ok (getf got-msg :success))
-      (ok (not (getf got-msg :timeout)))
-      (is (rmq-message-body inner-msg) test-msg)
-      (ok (getf (ack-message *node* inner-msg) :success)))))
+
+    (let ((got-msg (get-message *node*)))
+      (is (rmq-message-body got-msg) test-msg)
+      (is (nack-message *node* got-msg nil) :amqp-status-ok))
+    (is-error (get-message *node*) 'rabbitmq-library-error)))
 
 (subtest "nack works as expected (no requeue)"
   (let ((test-msg (format nil "test-~d" (get-universal-time))))
-    (ok (getf (send-message *node* *source-queue* test-msg) :success))
+    (is (send-message *checking-node* *source-queue* test-msg) :amqp-status-ok)
     (sleep .1)
-    (let* ((got-msg (get-message *node*))
-           (inner-msg (getf got-msg :result)))
-      (ok (getf got-msg :success))
-      (ok (not (getf got-msg :timeout)))
-      (is (rmq-message-body inner-msg) test-msg)
-      (ok (getf (nack-message *node* inner-msg nil) :success)))
-    (sleep .1)
-    (let* ((got-msg (get-message *node*)))
-      (ok (getf got-msg :success))
-      (ok (getf got-msg :timeout))
-      (is (getf got-msg :items) nil))))
+
+    (let ((got-msg (get-message *node*)))
+      (is (rmq-message-body got-msg) test-msg)
+      (is (nack-message *node* got-msg nil) :amqp-status-ok))
+    (is-error (get-message *node*) 'rabbitmq-library-error)))
 
 (shutdown *node*)
-(setf *node* (make-rmq-node #'(lambda (x) (format nil "test ~a" x))
-                            (format nil "test-rmq-node-~d" (get-universal-time))
-                            *conn* 1 *source-queue* *dest-queue* *fail-queue* :batch-size 10))
+(setf *node*
+      (make-rmq-node #'(lambda (x) (format nil "test ~a" x))
+                     (format nil "test-rmq-node-~d" (get-universal-time))
+                     *source-queue* *dest-queue* *fail-queue*
+                     :batch-size 10 :host (uiop:getenv "TEST_RMQ")))
 (startup *node* nil)
 
 (subtest "pull-messages-gets-full-batch"
   (iter:iterate
-     (iter:repeat 10)
-     (send-message *node* *source-queue* "testing"))
+    (iter:repeat 10)
+    (send-message *checking-node* *source-queue* "testing"))
+  (sleep .1)
 
-   (sleep .1)
+  (let ((items (pull-items *node*)))
+    (is (length items) 10)
+    (iter:iterate
+      (iter:for item in items)
+      (ack-message *node* item))))
 
-   (let* ((result (pull-items *node*))
-          (items (getf result :items)))
-     (ok (getf result :success))
-     (is (length items) 10)
-     (iter:iterate
-       (iter:for item in items)
-       (ack-message *node* item))))
 
 (subtest "pull-messages-gets-partial"
   (iter:iterate
     (iter:repeat 5)
-    (send-message *node* *source-queue* "testing"))
-
+    (send-message *checking-node* *source-queue* "testing"))
   (sleep .1)
 
-  (let* ((result (pull-items *node*))
-         (items (getf result :items)))
-    (ok (getf result :success))
+  (let ((items (pull-items *node*)))
     (is (length items) 5)
     (iter:iterate
       (iter:for item in items)
       (ack-message *node* item))))
 
-(subtest "pull-messages-handles-error-list"
-  (with-mocks ()
-    (answer get-message '(:error "test"))
-    (let ((result (pull-items *node*)))
-      (ok (not (getf result :success)))
-      (is (getf result :error) "test"))))
-
 (subtest "transform-items success"
   (let ((items `("1" "2" "3" "4" "5" "6" "7" "8" "9" "10")))
-
     (iter:iterate
       (iter:for item in items)
-      (send-message *node* *source-queue* item))
-
+      (send-message *checking-node* *source-queue* item))
     (sleep .1)
 
-    (let ((got-items (pull-items *node*)))
-      (ok (getf got-items :success))
-
-      (let ((new-items (transform-items *node* got-items)))
-        (ok (getf new-items :success))
-        (is (length (getf new-items :items)) 10)
-
-        (iter:iterate
-          (iter:for expect in items)
-          (iter:for got in (getf new-items :items))
-          (is (rmq-message-body got) (format nil "test ~a" expect))
-          (ack-message *node* got))))))
+    (let ((new-items (transform-items *node* (pull-items *node*))))
+      (is (length new-items) 10)
+      (iter:iterate
+        (iter:for expect in items)
+        (iter:for got in new-items)
+        (is (rmq-message-body got) (format nil "test ~a" expect))
+        (ack-message *node* got)))))
 
 (shutdown *node*)
-(setf *node* (make-rmq-node #'(lambda (x) (declare (ignore x)) (error "test"))
-                            (format nil "test-rmq-node-~d" (get-universal-time))
-                            *conn* 1 *source-queue* *dest-queue* *fail-queue* :batch-size 10))
+(setf *node*
+      (make-rmq-node #'(lambda (x) (declare (ignore x)) (error "test"))
+                     (format nil "test-rmq-node-~d" (get-universal-time))
+                     *source-queue* *dest-queue* *fail-queue*
+                     :batch-size 10 :host (uiop:getenv "TEST_RMQ")))
 (startup *node* nil)
 
 (subtest "transform-items failure"
   (let ((items `("1" "2" "3" "4" "5" "6" "7" "8" "9" "10")))
-
     (iter:iterate
       (iter:for item in items)
-      (send-message *node* *source-queue* item))
-
+      (send-message *checking-node* *source-queue* item))
     (sleep .1)
 
     (let ((got-items (pull-items *node*)))
-      (ok (getf got-items :success))
-
-      (let ((new-items (transform-items *node* got-items)))
-        (ok (not (getf new-items :success)))
-        (is (length (getf new-items :items)) 10)
-        (is-error (getf new-items :error) 'simple-error)
-
-        (iter:iterate
-          (iter:for expect in items)
-          (iter:for got in (getf new-items :items))
-          (is (rmq-message-body got) expect)
-          (ack-message *node* got))))))
+      (handler-case (transform-items *node* got-items)
+        (node-error (c)
+          (is (node-error/step c) :transform)
+          (is (length (node-error/items c)) (length items))
+          (iter:iterate
+            (iter:for expected in items)
+            (iter:for got in (node-error/items c))
+            (is (rmq-message-body got) expected)
+            (ack-message *node* got)))
+        (:no-error (res) (declare (ignore res))
+          (fail "transform should not have succeeded"))))))
 
 (shutdown *node*)
-(setf *node* (make-rmq-node #'(lambda (x) (format nil "test ~a" x))
-                            (format nil "test-rmq-node-~d" (get-universal-time))
-                            *conn* 1 *source-queue* *dest-queue* *fail-queue* :batch-size 10))
-(startup *node* nil)
+(shutdown *checking-node*)
+(setf *node*
+      (make-rmq-node #'(lambda (x) (format nil "test ~a" x))
+                     (format nil "test-rmq-node-~d" (get-universal-time))
+                     *source-queue* *dest-queue* *fail-queue*
+                     :batch-size 10 :host (uiop:getenv "TEST_RMQ")))
 
-(defvar *checking-node* (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
-                                       *conn* 2 *dest-queue* *source-queue* *fail-queue* :batch-size 10))
+(setf *checking-node*
+      (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
+                     *dest-queue* *source-queue* *fail-queue*
+                     :batch-size 10 :host (uiop:getenv "TEST_RMQ")))
+(startup *node* nil)
 (startup *checking-node* nil)
 
 (subtest "place items works"
   (let ((items `("1" "2" "3" "4" "5" "6" "7" "8" "9" "10")))
-
     (iter:iterate
       (iter:for item in items)
-      (send-message *node* *source-queue* item))
-
+      (send-message *checking-node* *source-queue* item))
     (sleep .1)
 
     (let ((first-got-items (pull-items *node*)))
-      (ok (getf first-got-items :success))
-      (ok (getf (place-items *node* first-got-items) :success))
-
+      (place-items *node* first-got-items)
       (sleep .1)
 
       (let ((second-got-items (pull-items *checking-node*)))
-        (ok (getf second-got-items :success))
-        (is (length (getf second-got-items :items)) (length (getf first-got-items :items)))
+        (is (length second-got-items) (length first-got-items))
         (iter:iterate
           (iter:for expected in items)
-          (iter:for got-first in (getf second-got-items :items))
-          (iter:for got-second in (getf first-got-items :items))
-          (string= expected (rmq-message-body got-first))
-          (string= expected (rmq-message-body got-second)))
+          (iter:for got-first in first-got-items)
+          (iter:for got-second in second-got-items)
+          (is expected (rmq-message-body got-first))
+          (is expected (rmq-message-body got-second))
+          (ack-message *checking-node* got-second))))))
 
-        (iter:iterate
-          (iter:for item in (getf second-got-items :items))
-          (ack-message *checking-node* item))))))
+(skip 1 "place items handles put failure")
 
-(subtest "place items handles put failure"
-  (let ((items `("1" "2" "3" "4" "5" "6" "7" "8" "9" "10")))
-
-    (iter:iterate
-      (iter:for item in items)
-      (send-message *node* *source-queue* item))
-
-    (sleep .1)
-
-    (let ((got-items (pull-items *node*)))
-      (ok (getf got-items :success))
-      (with-mocks ()
-        (answer send-message '(:success nil))
-        (let ((res (place-items *node* got-items)))
-          (ok (not (getf res :success)))
-          (is (getf res :items) (getf got-items :items))))
-
-      (iter:iterate
-        (iter:for item in (getf got-items :items))
-        (ack-message *node* item)))))
-
-(subtest "place items handles ack failure"
-  (let ((items `("1" "2" "3" "4" "5" "6" "7" "8" "9" "10")))
-
-    (iter:iterate
-      (iter:for item in items)
-      (send-message *node* *source-queue* item))
-
-    (sleep .1)
-
-    (let ((got-items (pull-items *node*)))
-      (ok (getf got-items :success))
-      (with-mocks ()
-        (answer send-message '(:success t))
-        (answer ack-message '(:success nil))
-        (let ((res (place-items *node* got-items)))
-          (ok (not (getf res :success)))
-          (is (getf res :items) (getf got-items :items))))
-
-      (iter:iterate
-        (iter:for item in (getf got-items :items))
-        (ack-message *node* item)))))
+(skip 1 "place items handles ack failure")
 
 (subtest "handle-failure-unexpected-step"
   (is-error (handle-failure *node* :bad nil) 'simple-error))
 
-(subtest "handle-failure-pull-step"
-  (is (handle-failure *node* :pull '(:error "test")) '(:error "test")))
-
 (shutdown *checking-node*)
-(setf *checking-node* (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
-                                     *conn* 2 *fail-queue* *dest-queue* *source-queue* :batch-size 10))
+(setf *checking-node*
+      (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
+                     *fail-queue* *dest-queue* *source-queue*
+                     :batch-size 10 :host (uiop:getenv "TEST_RMQ")))
 (startup *checking-node* nil)
+
+(subtest "handle-failure-pull-step send successful"
+  (iter:iterate
+    (iter:for item in '("1" "2" "3" "4" "5"))
+    (send-message *node* *source-queue* item))
+  (sleep .1)
+
+  (let ((got-items (pull-items *node*)))
+    (is (length got-items) 5)
+    (handle-failure *node* :pull got-items)
+    (sleep .1)
+
+    (let ((final-items (pull-items *checking-node*)))
+      (is (length final-items) (length got-items))
+      (iter:iterate
+        (iter:for got-item in got-items)
+        (iter:for final-item in final-items)
+        (is (rmq-message-body final-item) (rmq-message-body got-item))
+        (ack-message *checking-node* final-item)))))
+
+(skip 1 "handle-failure-pull-step-send-unsuccessful")
 
 (subtest "handle-failure-transform-step-send-successful"
   (iter:iterate
     (iter:for item in '("1" "2" "3" "4" "5"))
     (send-message *node* *source-queue* item))
-
   (sleep .1)
 
   (let ((got-items (pull-items *node*)))
-    (ok (getf got-items :success))
-    (is (length (getf got-items :items)) 5)
-    (setf got-items (append '(:error "test") got-items))
-    (is (handle-failure *node* :transform got-items)
-        got-items)
-
+    (is (length got-items) 5)
+    (handle-failure *node* :transform got-items)
     (sleep .1)
 
     (let ((final-items (pull-items *checking-node*)))
-      (ok (getf final-items :success))
-      (ok (not (getf final-items :timeout)))
-      (is (length (getf final-items :items)) (length (getf got-items :items)))
-
+      (is (length final-items) (length got-items))
       (iter:iterate
-        (iter:for got-item in (getf got-items :items))
-        (iter:for final-item in (getf final-items :items))
+        (iter:for got-item in got-items)
+        (iter:for final-item in final-items)
         (is (rmq-message-body final-item) (rmq-message-body got-item))
         (ack-message *checking-node* final-item)))))
 
-(subtest "handle-failure-transform-step-send-unsuccessful"
-  (iter:iterate
-    (iter:for item in '("1" "2" "3" "4" "5"))
-    (send-message *node* *source-queue* item))
-
-  (sleep .1)
-
-  (let ((got-items (pull-items *node*)))
-    (ok (getf got-items :success))
-    (is (length (getf got-items :items)) 5)
-    (setf got-items (append '(:error "test") got-items))
-    (with-mocks ()
-      (answer send-message '(:error "test"))
-      (is (handle-failure *node* :transform got-items)
-          got-items))
-
-    (sleep .1)
-
-    (let ((final-items (pull-items *node*)))
-      (ok (getf final-items :success))
-      (ok (not (getf final-items :timeout)))
-      (is (length (getf final-items :items)) (length (getf got-items :items)))
-
-      (iter:iterate
-        (iter:for got-item in (getf got-items :items))
-        (iter:for final-item in (getf final-items :items))
-        (is (rmq-message-body final-item) (rmq-message-body got-item))
-        (ack-message *node* final-item)))))
+(skip 1 "handle-failure-transform-step-send-unsuccessful")
 
 (subtest "handle-failure-place-step-send-successful"
   (iter:iterate
     (iter:for item in '("1" "2" "3" "4" "5"))
     (send-message *node* *source-queue* item))
-
   (sleep .1)
 
   (let ((got-items (pull-items *node*)))
-    (ok (getf got-items :success))
-    (is (length (getf got-items :items)) 5)
-    (setf got-items (append '(:error "test") got-items))
-    (is (handle-failure *node* :place got-items)
-        got-items)
-
+    (is (length got-items) 5)
+    (handle-failure *node* :place got-items)
     (sleep .1)
 
     (let ((final-items (pull-items *checking-node*)))
-      (ok (getf final-items :success))
-      (ok (not (getf final-items :timeout)))
-      (is (length (getf final-items :items)) (length (getf got-items :items)))
-
+      (is (length final-items) (length got-items))
       (iter:iterate
-        (iter:for got-item in (getf got-items :items))
-        (iter:for final-item in (getf final-items :items))
+        (iter:for got-item in got-items)
+        (iter:for final-item in final-items)
         (is (rmq-message-body final-item) (rmq-message-body got-item))
         (ack-message *checking-node* final-item)))))
 
-(subtest "handle-failure-place-step-send-unsuccessful"
-  (iter:iterate
-    (iter:for item in '("1" "2" "3" "4" "5"))
-    (send-message *node* *source-queue* item))
-
-  (sleep .1)
-
-  (let ((got-items (pull-items *node*)))
-    (ok (getf got-items :success))
-    (is (length (getf got-items :items)) 5)
-    (setf got-items (append '(:error "test") got-items))
-    (with-mocks ()
-      (answer send-message '(:error "test"))
-      (is (handle-failure *node* :place got-items)
-          got-items))
-
-    (sleep .1)
-
-    (let ((final-items (pull-items *node*)))
-      (ok (getf final-items :success))
-      (ok (not (getf final-items :timeout)))
-      (is (length (getf final-items :items)) (length (getf got-items :items)))
-
-      (iter:iterate
-        (iter:for got-item in (getf got-items :items))
-        (iter:for final-item in (getf final-items :items))
-        (is (rmq-message-body final-item) (rmq-message-body got-item))
-        (ack-message *node* final-item)))))
+(skip 1 "handle-failure-place-step-send-unsuccessful")
 
 (shutdown *node*)
 (shutdown *checking-node*)
-(setf *node* (make-rmq-node #'(lambda (x) (format nil "test ~a" x))
-                            (format nil "test-rmq-node-~d" (get-universal-time))
-                            *conn* 1 *source-queue* *dest-queue* *fail-queue* :batch-size 1))
-(setf *checking-node* (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
-                                     *conn* 2 *dest-queue* *source-queue* *fail-queue* :batch-size 5))
+(setf *node*
+      (make-rmq-node #'(lambda (x) (format nil "test ~a" x))
+                     (format nil "test-rmq-node-~d" (get-universal-time))
+                     *source-queue* *dest-queue* *fail-queue*
+                     :batch-size 1 :host (uiop:getenv "TEST_RMQ")))
+(setf *checking-node*
+      (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
+                     *dest-queue* *source-queue* *fail-queue*
+                     :batch-size 5 :host (uiop:getenv "TEST_RMQ")))
 (startup *node* nil)
 (startup *checking-node* nil)
 
@@ -379,60 +260,59 @@
   (let ((test-items '("1" "2" "3" "4" "5")))
     (iter:iterate
       (iter:for item in test-items)
-      (send-message *node* *source-queue* item))
-
+      (send-message *checking-node* *source-queue* item))
     (sleep .1)
 
     (iter:iterate
       (iter:repeat 5)
-      (ok (getf (run-iteration *node*) :success)))
-
+      (run-iteration *node*))
     (sleep .1)
 
-    (let* ((got-items (pull-items *checking-node*))
-           (inner-got-items (getf got-items :items)))
-      (ok (getf got-items :success))
+    (let* ((got-items (pull-items *checking-node*)))
+      (is (length test-items) (length got-items))
       (iter:iterate
         (iter:for test-item in test-items)
-        (iter:for got-item in inner-got-items)
+        (iter:for got-item in got-items)
         (is (format nil "test ~a" test-item) (rmq-message-body got-item))
         (ack-message *checking-node* got-item)))))
 
-(subtest "full node path - pull fail"
-  (let ((res '(:error "test")))
-    (with-mocks ()
-      (answer pull-items res)
-      (is (run-iteration *node*) res))))
-
 (shutdown *checking-node*)
-(setf *checking-node* (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
-                                     *conn* 2 *fail-queue* *source-queue* *dest-queue* :batch-size 5))
+(setf *checking-node*
+      (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
+                     *fail-queue* *source-queue* *dest-queue*
+                     :batch-size 5 :host (uiop:getenv "TEST_RMQ")))
 (startup *checking-node* nil)
+
+(subtest "full node path - pull fail"
+  (iter:iterate
+    (iter:repeat 5)
+    (with-mocks ()
+      (answer (pull-items _)
+        (error 'node-error :message "test" :step :pull))
+      (run-iteration *node*)))
+
+  (is (pull-items *checking-node*) nil))
 
 (subtest "full node path - transform fail"
   (let ((test-items '("1" "2" "3" "4" "5")))
     (iter:iterate
       (iter:for item in test-items)
-      (send-message *node* *source-queue* item))
-
+      (send-message *checking-node* *source-queue* item))
     (sleep .1)
 
     (iter:iterate
       (iter:repeat 5)
       (with-mocks ()
         (answer (transform-items _ items)
-          `(:error "test" :items ,(getf items :items)))
-        (ok (not (getf (run-iteration *node*) :success)))))
-
+          (error 'node-error :message "test" :items items :step :transform))
+        (run-iteration *node*)))
     (sleep .1)
 
-    (let* ((got-items (pull-items *checking-node*))
-           (inner-got-items (getf got-items :items)))
-      (ok (getf got-items :success))
-      (is (length inner-got-items) (length test-items))
+    (let ((got-items (pull-items *checking-node*)))
+      (is (length got-items) (length test-items))
       (iter:iterate
         (iter:for test-item in test-items)
-        (iter:for got-item in inner-got-items)
+        (iter:for got-item in got-items)
         (is test-item (rmq-message-body got-item))
         (ack-message *checking-node* got-item)))))
 
@@ -440,34 +320,73 @@
   (let ((test-items '("1" "2" "3" "4" "5")))
     (iter:iterate
       (iter:for item in test-items)
-      (send-message *node* *source-queue* item))
-
+      (send-message *checking-node* *source-queue* item))
     (sleep .1)
 
     (iter:iterate
       (iter:repeat 5)
       (with-mocks ()
         (answer (place-items _ items)
-          `(:error "test" :items ,(getf items :items)))
-        (ok (not (getf (run-iteration *node*) :success)))))
-
+          (error 'node-error :message "test" :items items :step :place))
+        (run-iteration *node*)))
     (sleep .1)
 
-    (let* ((got-items (pull-items *checking-node*))
-           (inner-got-items (getf got-items :items)))
-      (ok (getf got-items :success))
-      (is (length inner-got-items) (length test-items))
+    (let ((got-items (pull-items *checking-node*)))
+      (is (length got-items) (length test-items))
       (iter:iterate
         (iter:for test-item in test-items)
-        (iter:for got-item in inner-got-items)
+        (iter:for got-item in got-items)
         (is (format nil "test ~a" test-item) (rmq-message-body got-item))
         (ack-message *checking-node* got-item)))))
 
-(delete-queue *node* *source-queue*)
-(delete-queue *node* *dest-queue*)
-(delete-queue *node* *fail-queue*)
-(shutdown *node*)
 (shutdown *checking-node*)
-(destroy-connection *conn*)
+(defparameter *final-queue* (format nil "test-final-~d" (get-universal-time)))
+(defvar *second-node*
+  (make-rmq-node #'(lambda (x) (format nil "test1 ~a" x))
+                 (format nil "test-rmq-node-2-~d" (get-universal-time))
+                 *dest-queue* *final-queue* *fail-queue*
+                 :batch-size 1 :host (uiop:getenv "TEST_RMQ")))
+(setf *checking-node*
+      (make-rmq-node nil (format nil "test-rmq-node-1-~d" (get-universal-time))
+                     *final-queue* *source-queue* *dest-queue*
+                     :batch-size 5 :host (uiop:getenv "TEST_RMQ")))
+(startup *second-node* nil)
+(startup *checking-node* nil)
+
+(subtest "full node path - success - two nodes"
+  (let ((test-items '("1" "2" "3" "4" "5")))
+    (iter:iterate
+      (iter:for item in test-items)
+      (send-message *checking-node* *source-queue* item))
+    (sleep .1)
+
+    (iter:iterate
+      (iter:repeat 5)
+      (iter:for i upfrom 0)
+      (run-iteration *node*)
+      (diag (format nil "first iter: ~d" i))
+      (sleep .1)
+      (run-iteration *second-node*)
+      (diag (format nil "second iter: ~d" i))
+      (sleep .1))
+
+    (let ((got-items (pull-items *checking-node*)))
+      (is (length got-items) (length test-items))
+      (iter:iterate
+        (iter:for test-item in test-items)
+        (iter:for got-item in got-items)
+        (is (format nil "test1 test ~a" test-item) (rmq-message-body got-item))
+        (ack-message *checking-node* got-item)))))
+
+(defvar delete-conn (setup-connection :host (uiop:getenv "TEST_RMQ")))
+(with-channel (delete-conn 1)
+  (queue-delete delete-conn 1 *source-queue*)
+  (queue-delete delete-conn 1 *dest-queue*)
+  (queue-delete delete-conn 1 *final-queue*)
+  (queue-delete delete-conn 1 *fail-queue*))
+(destroy-connection delete-conn)
+(shutdown *node*)
+(shutdown *second-node*)
+(shutdown *checking-node*)
 
 (finalize)