+++ CHANGELOG.md
@@ -16,4 +16,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added MMOP/0 start-node-failure.
 - Added MMOP/0 stop-worker.
 - Added node recipe system for constructing node threads.
-- rmq-worker startup/shutdown
+- rmq-worker startup/shutdown, node start up, and event loop.